RamiLab2

+++ src/RamiLab2.cpp
@@ -10,7 +10,7 @@ void loop();
 SYSTEM_MODE(MANUAL);
 SYSTEM_THREAD(ENABLED);
 
-uint16_t value; 
+uint8_t value; 
 
 
 void setup() {
