+++ src/RamiLab2.ino
@@ -1,7 +1,7 @@
 SYSTEM_MODE(MANUAL);
 SYSTEM_THREAD(ENABLED);
 
-uint16_t value; 
+uint8_t value; 
 
 
 void setup() {
